feat: login

+++ src/api/api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import API_CONSTANTS, { BASE_URL, TOKEN } from "./constants";
+import API_CONSTANTS, { BASE_URL, TOKEN_LOCAL_STORAGE } from "./constants";
 import { useNavigate } from "react-router-dom";
 
 const api = axios.create({
@@ -13,23 +13,21 @@ const api = axios.create({
 });
 
 api.interceptors.request.use(async (config) => {
-  //   if (
-  //     config.url === API_CONSTANTS.AUTH.SIGNUP_CUSTOMER ||
-  //     config.url === API_CONSTANTS.AUTH.LOGIN_CUSTOMER ||
-  //     config.url === API_CONSTANTS.AUTH.LOGIN_STAFF
-  //   ) {
-  //     console.log("okvcl");
-  //   } else
-
-  const navigate = useNavigate();
-  try {
-    const token = JSON.parse(localStorage.getItem(TOKEN));
-    console.log("dcmtoken", token);
-    token && (config.headers[TOKEN] = `${token}`);
-  } catch (error) {
-    localStorage.clear();
-    navigate("/login");
-    throw error.response.data;
+  if (
+    config.url === API_CONSTANTS.AUTH.SIGNUP_CUSTOMER ||
+    config.url === API_CONSTANTS.AUTH.LOGIN_CUSTOMER ||
+    config.url === API_CONSTANTS.AUTH.LOGIN_STAFF
+  ) {
+  } else {
+    const navigate = useNavigate();
+    try {
+      const token = localStorage.getItem(TOKEN_LOCAL_STORAGE);
+      token && (config.headers[TOKEN_LOCAL_STORAGE] = `${token}`);
+    } catch (error) {
+      localStorage.clear();
+      navigate("/login");
+      throw error.response.data || error;
+    }
   }
 
   return config;
@@ -40,7 +38,7 @@ api.interceptors.response.use(
     return response.data;
   },
   (error) => {
-    throw error.response.data;
+    throw error.response.data || error;
   }
 );
